requirements.txt

sentence-transformers==2.2.2
# huggingface-hub<0.22.0
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]
langchain-pinecone
langchain-community
langchain-groq
langchain-experimental
-e .